Improve enum class

+++ src/STM32Serial.cpp
@@ -25,7 +25,7 @@ void STM32Serial::init( uint32_t baudrate, LengthSerial databits, ParitySerial p
     serial.pin_rx = (PinName)rxPin.name;
     serial.pin_tx = (PinName)txPin.name;
 
-    uart_init( &serial, baudrate, databits, parity, stopBit );
+    uart_init( &serial, baudrate, (uint32_t)databits, (uint32_t)parity, (uint32_t)stopBit );
     uart_attach_rx_callback(&serial, RxIRQ);
 }
 
@@ -52,7 +52,7 @@ int STM32Serial::setBaudrate(uint32_t baudrate){
     this->baudrate = baudrate;
 
     uart_deinit(&serial);
-    uart_init(&serial, baudrate, databits, parity, stopBit);
+    uart_init(&serial, baudrate, (uint32_t)databits, (uint32_t)parity, (uint32_t)stopBit);
 
     return DEVICE_OK;
 }
@@ -66,7 +66,7 @@ int STM32Serial::configurePins(Pin& tx, Pin& rx){
     serial.pin_tx = (PinName)tx.name;
 
     uart_deinit(&serial);
-    uart_init(&serial, baudrate, databits, parity, stopBit);
+    uart_init(&serial, baudrate, (uint32_t)databits, (uint32_t)parity, (uint32_t)stopBit);
 
     return DEVICE_OK;
 }
